+++ api/internal/rooms/application/useCases/joinRoomUsecase.go
