+++ api/internal/rooms/application/useCases/websocket/manageUsecase.go
@@ -5,8 +5,6 @@ import (
 	"errors"
 	"fmt"
 	"log"
-	"strconv"
-	"sync"
 
 	"github.com/gorilla/websocket"
 
@@ -14,106 +12,6 @@ import (
 
 	sv "suffgo/internal/shared/domain/valueObjects"
 )
-
-// Message representa un mensaje que se difunde a los clientes conectados.
-type Message struct {
-	SenderID string
-	Data     []byte
-}
-
-type Client struct {
-	Username string
-	Conn     *websocket.Conn
-}
-
-// Hub administra las conexiones y la difusión de mensajes en una sala.
-type Hub struct {
-	Clients    map[string]*Client // Clave: Username
-	Broadcast  chan Message
-	Register   chan *Client
-	Unregister chan *Client
-}
-
-// NewHub crea e inicializa un nuevo Hub.
-func NewHub() *Hub {
-	return &Hub{
-		Clients:    make(map[string]*Client),
-		Broadcast:  make(chan Message),
-		Register:   make(chan *Client),
-		Unregister: make(chan *Client),
-	}
-}
-
-// Run inicia el bucle principal del Hub, gestionando registros, desregistros y difusión.
-func (h *Hub) Run() {
-	for {
-		select {
-		case client := <-h.Register:
-			// Si ya existe un cliente con ese UserID, cerramos la conexión anterior.
-			if oldClient, exists := h.Clients[client.Username]; exists {
-				log.Printf("El usuario %s ya estaba conectado, cerrando la conexión anterior.", client.Username)
-				oldClient.Conn.Close()
-			}
-			h.Clients[client.Username] = client
-			log.Printf("Cliente registrado: %s. Total clientes: %d", client.Username, len(h.Clients))
-
-		case client := <-h.Unregister:
-			if _, exists := h.Clients[client.Username]; exists {
-				delete(h.Clients, client.Username)
-				client.Conn.Close()
-				log.Printf("Cliente desregistrado: %s. Total clientes: %d", client.Username, len(h.Clients))
-			}
-
-		case msg := <-h.Broadcast:
-			// Se envía el mensaje a todos los clientes excepto al emisor.
-			for uid, client := range h.Clients {
-				if uid != msg.SenderID {
-					if err := client.Conn.WriteMessage(websocket.TextMessage, msg.Data); err != nil {
-						log.Printf("Error al enviar mensaje a %s: %v", uid, err)
-						client.Conn.Close()
-						delete(h.Clients, uid)
-					}
-				}
-			}
-		}
-	}
-}
-
-type RoomManager struct {
-	mu    sync.Mutex
-	Rooms map[string]*Hub
-}
-
-// NewRoomManager crea e inicializa un RoomManager.
-func NewRoomManager() *RoomManager {
-	return &RoomManager{
-		Rooms: make(map[string]*Hub),
-	}
-}
-
-func (rm *RoomManager) GetHub(roomID string) *Hub {
-	rm.mu.Lock()
-	defer rm.mu.Unlock()
-	hub, exists := rm.Rooms[roomID]
-	if !exists {
-		return nil
-	}
-	return hub
-}
-
-// Solo lo debe poder usar el administrador
-func (rm *RoomManager) InitializeHub(roomID uint) *Hub {
-	roomIDStr := strconv.FormatUint(uint64(roomID), 10)
-	hub := NewHub()
-	rm.Rooms[roomIDStr] = hub
-	go hub.Run()
-	return hub
-}
-
-type ClientAction struct {
-	Action  string          `json:"action"`
-	Payload json.RawMessage `json:"payload"`
-}
 
 type ManageWsUsecase struct {
 	repository domain.RoomRepository
@@ -166,7 +64,7 @@ func (s *ManageWsUsecase) Execute(ws *websocket.Conn, username, roomID string, c
 		response := fmt.Sprintf("%s: %s", username, payload.Message)
 
 		hub.Broadcast <- Message{SenderID: username, Data: []byte(response)}
-	case "start_room":
+	case "start_room": 
 		//Obtengo la sala a partir del id
 
         roomIDobj, err := sv.NewID(roomID)
@@ -200,6 +98,7 @@ func (s *ManageWsUsecase) Execute(ws *websocket.Conn, username, roomID string, c
 
 
 		//falta implementar go routine que contiene bucle de lectura 
+		go hub.Run()
 	case "join_room":
 	default:
 		log.Printf("Acción desconocida: %s", clientAction.Action)
@@ -207,6 +106,8 @@ func (s *ManageWsUsecase) Execute(ws *websocket.Conn, username, roomID string, c
 		if err := ws.WriteMessage(websocket.TextMessage, []byte(errMsg)); err != nil {
 			log.Printf("Error al enviar mensaje de error: %v", err)
 		}
+
+		
 	}
 	return nil
 }
